+++ src/Render/D3D12/D3D12Utils.h
@@ -14,7 +14,7 @@ namespace D3D12Utils
 	ID3DBlob* CompileBlobFromFile(const char* filePath, const char* entryName, const char* target, u32 flags);
 	ID3DBlob* CompileBlob(const void* date, const i32 dataSize, const char* entryName, const char* target, u32 flags);
 
-	ID3D12Resource* CreateUploadBuffer(ID3D12Device* device, u32 size, const char* name = nullptr);
+	ID3D12Resource* CreateUploadBuffer(ID3D12Device* device, u64 size, const char* name = nullptr);
 
 	void SetRawD3D12ResourceName(ID3D12Resource* res, const char* name);
 	void SetRawD3D12ResourceName(ID3D12Resource* res, const std::string& name);
